sqlx-postgres: Remove PgStream
